fix: render CSS correctly on both client and server
