+++ test/src/modules/resolver/scopedCss/basic.html
@@ -1,0 +1,9 @@
+<!--
+Copyright (c) 2018, salesforce.com, inc.
+All rights reserved.
+SPDX-License-Identifier: MIT
+For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/MIT
+-->
+<template>
+    <h1>Basic</h1>
+</template>
